done constantcolumn, add benchmark

+++ tests/funcexp_bench/test.sql
@@ -1,7 +1,7 @@
 use test;
 
 drop table if exists t1;
-create table t1 (a int, b int) engine=columnstore;
+create table t1 (a tinyint, b tinyint) engine=columnstore;
 
 insert into t1 values (1,1),(2,2),(3,3),(4,4),(5,5),(6,6),(7,7),(8,8),(9,9),(10,10);
